backend/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
